requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1
djangorestframework==3.11.1
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.6.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
